moved ApplicationAuditEvent into models package
