+++ packages/models/src/index.ts
@@ -1,7 +1,8 @@
 // Models
 export * from "./queue.js";
 export * from "./db.js";
-export * from "./events/events.js";
+export * from "./events.js";
+export * from "./applicationAudit.js";
 
 // Utilities
 export * from "./brandedIds.js";
